feat: Rename recurrence pages
- RecurrenceBundlePage -> RecurrencePage
- RecurrencePage -> RecurrencesPage
