+++ src/ducks/recurrence/index.js
@@ -1,3 +1,3 @@
 export { default as DebugRecurrencePage } from './DebugRecurrencePage'
+export { default as RecurrencesPage } from './RecurrencesPage'
 export { default as RecurrencePage } from './RecurrencePage'
-export { default as RecurrenceBundlePage } from './RecurrenceBundlePage'
